feat: better structure login screen
