+++ src/ui/mobile/routes/login.tsx
@@ -1,4 +1,4 @@
-import { LoginScreen } from "@/ui/mobile/modules/login/LoginScreen";
+import { LoginScreen } from "@/ui/mobile/modules/login/screens/LoginScreen";
 import { RequireUnAuth } from "@/ui/mobile/router/auth";
 
 export default function LoginRoute() {
